+++ resolve/src/util/network/mod.rs
@@ -1,0 +1,28 @@
+//! Utilities for networking.
+
+use std::net::{Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6};
+use std::task::Poll;
+
+pub mod http;
+pub mod http_async;
+pub mod retry;
+
+/// LOCALHOST constants for both IPv4 and IPv6.
+pub const LOCALHOST: [SocketAddr; 2] = [
+    SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 0)),
+    SocketAddr::V6(SocketAddrV6::new(Ipv6Addr::LOCALHOST, 0, 0, 0)),
+];
+
+pub trait PollExt<T> {
+    fn expect(self, msg: &str) -> T;
+}
+
+impl<T> PollExt<T> for Poll<T> {
+    #[track_caller]
+    fn expect(self, msg: &str) -> T {
+        match self {
+            Poll::Ready(val) => val,
+            Poll::Pending => panic!("{}", msg),
+        }
+    }
+}